requirements/test.txt
-r base.txt
factory-boy==2.8.1
Faker==0.7.10